WIP Routing working.
This remains a Work In Progress for several reasons, the idea being to clean what has been done.

- there are some little changes that can be moved to the code cleaning commit
- we should discuss about the configuration I made to simplify the paths to 3rd party modules installed by bower (using noder-js)
- the .gitignore to exclude unneeded stuff from bower components could be simplified, more by selective inclusion rather than selective exlusion (don't know if it's possible somehow)
- should I keep the minified version of director or not?

And most of all: simplify or enhance the filter specifications part. For now the design is a bit lame, I would rather turn the filterLink template into a component, with a proper controller. That would make more sense.